requirements.txt
amqp==2.5.2
asgiref==3.3.1
asn1crypto==1.3.0
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
awscli==1.18.69
base58==2.0.0
bcrypt==3.1.7
bencode.py==4.0.0
billiard==3.6.3.0
bit==0.6.1
boto3==1.13.19
botocore==1.16.19
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
channels==3.0.3
chardet==3.0.4
click==7.1.2
clinner==1.12.3
coincurve==13.0.0
colorama==0.4.3
colorlog==3.2.0
constantly==15.1.0
cryptography==3.4.6
cytoolz==0.10.1
daphne==3.0.1
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-cors-headers==3.3.0
django-health-check==3.12.1
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-simple-sso==0.14.1
django-storages==1.9.1
djangorestframework==3.11.0
dnslib==0.9.12
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
ecdsa==0.15
eciespy==0.3.5
entrypoints==0.3
eth-hash==0.2.0
eth-keys==0.3.3
eth-typing==2.2.1
eth-utils==1.9.0
flake8==3.8.2
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
gunicorn==20.0.4
h11==0.12.0
httptools==0.1.1
hyperlink==21.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
itsdangerous==1.1.0
jmespath==0.10.0
keylib==0.1.1
kombu==4.6.8
localstack==0.11.2
localstack-client==0.24
localstack-ext==0.11.17
mccabe==0.6.1
numpy==1.18.4
oauthlib==3.1.0
packaging==20.4
pip-review==1.1.0
psycopg2==2.8.5
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.7
pycryptodomex==3.9.8
pyflakes==2.2.0
PyHamcrest==2.0.2
pyjwkest==1.4.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.3.3
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
subprocess32==3.5.4
toolz==0.10.0
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
urllib3==1.25.9
utilitybelt==0.2.6
uvicorn==0.13.3
uvloop==0.15.1
vine==1.3.0
watchgod==0.6
webservices==0.7
websockets==8.1
zipp==3.1.0
zope.interface==5.2.0
